Factor out attacks and add to goblin
