+++ Assets/Scripts/Attacker.cs
@@ -1,0 +1,60 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Attacker : MonoBehaviour
+{
+    private bool isAttacking = false;
+    public bool IsAttacking { get => isAttacking; set => isAttacking = value; }
+
+    public int attackDamage = 1;
+
+    public Transform attackPoint;
+    public float attackRange = 0.5f;
+    public LayerMask enemyLayers;
+
+    Animator animator;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        animator = GetComponent<Animator>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void Attack()
+    {
+        IsAttacking = true;
+        animator.SetTrigger("Attack");
+
+        Collider2D[] hitEnemies = Physics2D.OverlapCircleAll(attackPoint.position, attackRange, enemyLayers);
+
+        foreach (Collider2D enemy in hitEnemies)
+        {
+            Damageable damageable = enemy.GetComponent<Damageable>();
+            damageable.ChangeHealth(-attackDamage);
+
+            Debug.Log("We hit " + enemy.name);
+        }
+    }
+
+    void OnDrawGizmosSelected()
+    {
+        if (attackPoint == null)
+        {
+            return;
+        }
+        Gizmos.DrawWireSphere(attackPoint.position, attackRange);
+    }
+
+    void CompleteAttack()
+    {
+        //Debug.Log("completed attack");
+        IsAttacking = false;
+    }
+}
