+++ Assets/Scripts/Attacker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8734c8cd94e8247c4a239552880f8f01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
