+++ Assets/Scripts/States/Patrol.cs
@@ -4,6 +4,7 @@ using UnityEngine.AI;
 public class Patrol : IState
 {
     private KnightController _knight;
+    private Attacker _attacker;
     private Vector2 _patrolDirection;
     private float planTimer;
     float patrolLength = 2.0f;
@@ -13,6 +14,7 @@ public class Patrol : IState
     {
         _knight = knight;
         _patrolDirection = patrolDirection.normalized;
+        _attacker = _knight.GetComponent<Attacker>();
     }
 
     public void OnEnter()
@@ -23,14 +25,14 @@ public class Patrol : IState
 
     public void Tick()
     {
-        if (!_knight.IsAttacking)
+        if (!_attacker.IsAttacking)
         {
             planTimer += Time.deltaTime;
             _knight.SetMoveDirection(_patrolDirection);
             if (planTimer > patrolLength)
             {
                 _patrolDirection *= -1.0f;
-                _knight.Attack();
+                _attacker.Attack();
                 planTimer = 0.0f;
             }
         }
